(1) Added more tests to tests/testReconfigurationMain/testReconfigurationMain.sh to include addition and deletion of active replica or reconfigurator nodes.
(2) Made non-backwards-compatible, compile breaking changes to
AbstractReplicaCoordinator.coordinateRequest and the ReplicableRequest
interface that should be straightforward to fix for gigapaxos
applications. The main change is to introduce cleaner support for
callbacks and, for that purpose, enforce a long request identifier and
a corresponding ReplicableRequest.getRequestID() method in
ReplicableRequest. Requests already also implementing the
ClientRequest interface should work unmodified.